+++ invoiceflow/src/main/java/com/invoiceflow/branding/UserLogoRepository.java
@@ -1,0 +1,5 @@
+package com.invoiceflow.branding;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserLogoRepository extends JpaRepository<UserLogo, Long> {}